Conflicts:
	bin/game/Game.class
	bin/game/parts/Actionbar.class
	bin/game/parts/GameMenu.class
	bin/game/parts/Inventory.class
	src/game/Game.java
	src/game/parts/Actionbar.java
	src/game/parts/GameMenu.java
	src/game/parts/Inventory.java

+++ src/worldObjects/food/ingredients/GIngredient.java
@@ -10,7 +10,7 @@ public abstract class GIngredient extends GFood{
 
 	boolean heldByMouse;
 	
-	public GIngredient(float xPos, float yPos, float texSize, float weight, float headPos, float footPos, float health) {
+	public GIngredient(float xPos, float yPos, float texSize, float weight, float headPos, float footPos, int health) {
 		super( xPos,  yPos,  texSize,  weight,  headPos,  footPos, health);
 		setTexFolder("food/ingredients/");
 	}
